Move all the inline code from the `index.html` to the JS files.  Rename the `load-src.js` to `window.js` to better reflect the contents of the file.  Display version number in the window title bar.
